pyproject.toml
[project]
name = "rtvc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.2.1",
    "fastapi[standard]>=0.115.6",
    "pydantic>=2.10.5",
    "twilio>=9.4.3",
    "uvicorn>=0.34.0",
    "websockets>=14.2",
]

[dependency-groups]
dev = [
    "ipython>=8.31.0",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.6",
]
